Fix #259 Move classes used solely by download mojos to org.codehaus.mojo.license.download
