+++ src/test/java/org/codehaus/mojo/license/download/LicenseSummaryTest.java
@@ -1,31 +1,8 @@
-package org.codehaus.mojo.license.utils;
+package org.codehaus.mojo.license.download;
 
 import org.codehaus.mojo.license.Eol;
-import org.codehaus.mojo.license.model.ProjectLicense;
-
-/*
- * #%L
- * License Maven Plugin
- * %%
- * Copyright (C) 2008 - 2011 CodeLutin, Codehaus, Tony Chemit
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Lesser Public License for more details.
- *
- * You should have received a copy of the GNU General Lesser Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/lgpl-3.0.html>.
- * #L%
- */
-
-import org.codehaus.mojo.license.model.ProjectLicenseInfo;
+import org.codehaus.mojo.license.download.LicenseSummaryReader;
+import org.codehaus.mojo.license.download.LicenseSummaryWriter;
 import org.junit.Assert;
 import org.junit.Test;
 import org.xml.sax.SAXException;
